+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.9.7](https://github.com/introspection-org/pi-recipes/compare/v0.9.6...v0.9.7) (2026-07-15)
+
+
+### Bug Fixes
+
+* export MCP catalog preloading ([#83](https://github.com/introspection-org/pi-recipes/issues/83)) ([cbfe168](https://github.com/introspection-org/pi-recipes/commit/cbfe1681e4637005e27e54e2e293c4fda66a5121))
+
+## [0.9.6](https://github.com/introspection-org/pi-recipes/compare/v0.9.5...v0.9.6) (2026-07-15)
+
+
+### Bug Fixes
+
+* warm MCP catalogs asynchronously ([#81](https://github.com/introspection-org/pi-recipes/issues/81)) ([b21a500](https://github.com/introspection-org/pi-recipes/commit/b21a50067e448f514ab39f95b83e6b1be6ac6881))
+
 ## [0.9.5](https://github.com/introspection-org/pi-recipes/compare/v0.9.4...v0.9.5) (2026-07-15)